Remove CallingHttpClient.kt
